Ajout de la gestion des tentatives sur la page de chargement Ajout du namespace background pour le fond de terrain Ajout des class de gestion du fond début d'implémentation du dessin de fond grille Début d'implémentation de récupération de configuration et dessin du fond grille changement de manière d'afficher les logs (-> TraceLog) changement du style d'indentation du code

+++ src/game/scene/scene/BaseScene.hpp
@@ -101,12 +101,12 @@ namespace Game::Scene::Scene{
             /**
              * largeur de la fenêtre
              */
-            int width;
+            float width;
 
             /**
              * hauteur de la fenêtre
              */
-            int height;
+            float height;
 
             /**
              * nom de la fenêtre
